+++ knowde/complex/auth/cli/test_proc.py
@@ -4,7 +4,7 @@
 from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
-from knowde.complex.auth.routers import auth_router
+from knowde.complex.auth.routers import auth_router, user_router
 
 from .proc import (
     change_me_proc,
@@ -18,6 +18,7 @@ def _api_client() -> TestClient:
     """Test util."""
     api = FastAPI()
     api.include_router(auth_router)
+    api.include_router(user_router)
     return TestClient(api)
 
 
@@ -34,7 +35,8 @@ def test_crud_user() -> None:
     assert login_proc(email, password, post=client.post).is_success
 
     res = get_me_proc(client.get)
-    email2 = "user@gmail.com"
+    assert res.is_success
     assert res.json()["email"] == email
+    email2 = "user@gmail.com"
     res = change_me_proc(email2, patch=client.patch)
     assert res.json()["email"] == email2
