+++ knowde/complex/auth/errors.py
@@ -1,0 +1,5 @@
+"""auth errors."""
+
+
+class TokenUnsavedError(Exception):
+    """認証用トークンが保存されていない."""
