+++ knowde/feature/api/__init__.py
@@ -7,7 +7,7 @@ from typing import AsyncGenerator
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from knowde.complex.auth.routers import auth_router
+from knowde.complex.auth.routers import auth_router, user_router
 from knowde.feature.api.middle import Neo4jTransactionMiddleware, neo4j_logger
 from knowde.primitive.__core__ import ErrorHandlingMiddleware
 from knowde.primitive.config.env import Settings
@@ -44,6 +44,7 @@ api.include_router(def_router)
 api.include_router(p_router)
 api.include_router(deduct_router)
 api.include_router(auth_router)
+api.include_router(user_router)
 
 
 @api.get("/health")
